requirements.txt
requests==2.31.0
praw==7.7.1
openai==0.28.0
python-dotenv==1.0.0
discord.py==2.3.2
schedule==1.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
